Rename from 'InnerPopoverContent' to 'AccountUserOptions'
