+++ src/ui/SideBar/AccountUser/AccountUser.tsx
@@ -6,7 +6,7 @@ import { useMe } from 'client'
 
 import { useIsOpen } from '../useSideBar'
 
-import { InnerPopoverContent } from './InnerPopoverContent'
+import { AccountUserOptions } from './AccountUserOptions'
 
 import * as S from './AccountUser.styled'
 
@@ -22,7 +22,7 @@ export function AccountUser() {
     <Popover
       name={user?.like_be_called}
       customWidth='20.2rem'
-      innerContent={<InnerPopoverContent />}
+      innerContent={<AccountUserOptions />}
       wrapperStyle={{ marginTop: 'auto' }}
       reajustOnChangeSize
     >
