+++ src/ui/SideBar/AccountUser/AccountUserOptions.tsx
@@ -4,33 +4,33 @@ import { AiOutlinePieChart, AiOutlineLogout } from 'react-icons/ai'
 
 import { LogOut } from './LogOut'
 
-import * as S from './InnerPopoverContent.styled'
+import * as S from './AccountUserOptions.styled'
 
-export function InnerPopoverContent() {
+export function AccountUserOptions() {
   const [isLogOutVisible, setIsLogOutVisible] = useState(false)
 
   return (
     <S.Container>
       <li>
-        <button>
+        <button aria-label='profile'>
           <FiSmile size={20} />
           Profile
         </button>
       </li>
       <li>
-        <button>
+        <button aria-label='progress'>
           <AiOutlinePieChart size={20} />
           Progress
         </button>
       </li>
       <li>
-        <button>
+        <button aria-label='sessions'>
           <FiActivity size={20} />
           Sessions
         </button>
       </li>
       <li>
-        <button onClick={() => setIsLogOutVisible(true)}>
+        <button aria-label='log out' onClick={() => setIsLogOutVisible(true)}>
           <AiOutlineLogout size={20} />
           Log out
         </button>
